separated Status into a separate module as well
